requirements.txt
grpcio==1.11.0
grpcio-tools==1.11.0
Pillow==5.1.0
